+++ app/controllers/services_controller.rb
@@ -11,6 +11,6 @@ class ServicesController < ApplicationController
 
   def show
     @service = Service.find(params[:id])
-    @service_opinions = ServiceOpinion.joins(:order).where(orders: { service: @service })
+    @service_opinions = ServiceOpinion.joins(:project_item).where(project_items: { service: @service })
   end
 end