+++ config/routes.rb
@@ -12,8 +12,8 @@ Rails.application.routes.draw do
 
   resources :services, only: [:index, :show]
   resources :categories, only: :show
-  resources :orders, only: [:index, :show, :new, :create] do
-    scope module: :orders do
+  resources :project_items, only: [:index, :show, :new, :create] do
+    scope module: :project_items do
       resources :questions, only: [:index, :create]
       resources :service_opinions, only: [:new, :create]
     end