+++ db/migrate/20180716124933_create_service_opinions.rb
@@ -3,7 +3,7 @@ class CreateServiceOpinions < ActiveRecord::Migration[5.2]
     create_table :service_opinions do |t|
       t.integer :rating, null: false
       t.text :opinion
-      t.belongs_to :order, null: false, index: true
+      t.belongs_to :project_item, null: false, index: true
 
       t.timestamps
     end