+++ spec/factories/order_changes.rb
@@ -3,7 +3,7 @@
 FactoryBot.define do
   factory :order_change do
     status :created
-    sequence(:message) { |n| "order change #{n} message" }
-    order
+    sequence(:message) { |n| "project_item change #{n} message" }
+    project_item
   end
 end