+++ spec/factories/orders.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-FactoryBot.define do
-  factory :order do
-    status :created
-    user
-    service
-  end
-end