+++ spec/models/service_spec.rb
@@ -27,7 +27,7 @@ RSpec.describe Service do
 
   it { should have_many(:service_categories).dependent(:destroy) }
   it { should have_many(:categories) }
-  it { should have_many(:orders) }
+  it { should have_many(:project_items) }
 
   it "sets first category as default" do
     c1, c2 = create_list(:category, 2)